Performance and usability improvements
Resolved a CPU usage issue, dropping usage in worst case from 50% to 25% after switching from swapBuffers() to glFlush(). The other 25% appears to be related to the glUniform*() calls but I'm unsure how I can speed those up.

Fixed zooming to be more linear when zooming out. Previously, zooming out didn't actually change the same as the previous zoom inward did. This messed up the zoom level number completely. I tried to think of a nice solution in maths but today isn't my day. I ended up just implementing a zoom stack and on unzoom we just pop the last zoom level and set to that. Works fine.

I added in a 32-bit version of the fragment shader just to compare the quality of varying zoom levels. The same difference is expected when I get >fp64 implemented. You can switch current FP modes using the 1 and 2 keys on the numpad.

Finally I added in 3 location bookmarks for convenience. Using the 1, 2, and 3 keys on the keyboard will move the view to those locations.

+++ src/mandelbrot_fp32.frag
@@ -5,10 +5,10 @@ layout(location = 0) uniform int WINDOW_SIZE_WIDTH;
 layout(location = 1) uniform int WINDOW_SIZE_HEIGHT;
 layout(location = 2) uniform int CURRENT_COLOR_MODE;
 layout(location = 3) uniform int ESCAPE_VELOCITY_TEST_ITERATIONS;
-layout(location = 4) uniform double ORTHO_WIDTH;
-layout(location = 5) uniform double ORTHO_HEIGHT;
-layout(location = 6) uniform double BOUND_LEFT;
-layout(location = 7) uniform double BOUND_BOTTOM;
+layout(location = 4) uniform float ORTHO_WIDTH;
+layout(location = 5) uniform float ORTHO_HEIGHT;
+layout(location = 6) uniform float BOUND_LEFT;
+layout(location = 7) uniform float BOUND_BOTTOM;
 
 layout(location = 0) out vec4 fragColor;
 
@@ -22,23 +22,23 @@ vec3 hsv2rgb(vec3 c)
 }
 
 
-dvec2 complexAdd(dvec2 c1, dvec2 c2) {
-    return dvec2(c1.x+c2.x, c1.y+c2.y);
+vec2 complexAdd(vec2 c1, vec2 c2) {
+    return vec2(c1.x+c2.x, c1.y+c2.y);
 }
-dvec2 complexMulti(dvec2 c1, dvec2 c2) {
-    double imag = (c1.x * c2.y) + (c1.y * c2.x);
-    double real = (c1.x * c2.x) + (c1.y * c2.y * -1.0);
-    return dvec2(real,imag);
+vec2 complexMulti(vec2 c1, vec2 c2) {
+    float imag = (c1.x * c2.y) + (c1.y * c2.x);
+    float real = (c1.x * c2.x) + (c1.y * c2.y * -1.0);
+    return vec2(real,imag);
 }
-double complexMagnitude(dvec2 c) {
+float complexMagnitude(vec2 c) {
     if (c.x == 0.0 && c.y == 0.0) {
         return 0.0;
     } else {
         return sqrt(c.x*c.x + c.y*c.y);
     }
 }
-float findEscapeVelocity(dvec2 c) {
-    dvec2 z = dvec2(0.0, 0.0);
+float findEscapeVelocity(vec2 c) {
+    vec2 z = vec2(0.0, 0.0);
     int iter = 1;
     while (iter < ESCAPE_VELOCITY_TEST_ITERATIONS) {
         if (complexMagnitude(z) > 2.0) {
@@ -50,10 +50,10 @@ float findEscapeVelocity(dvec2 c) {
     }
     return 0.0;
 }
-dvec2 getOrthoCoordsFromWindowCoords(double x, double y) {
-    double orthoX = (x/double(WINDOW_SIZE_WIDTH) * ORTHO_WIDTH) + BOUND_LEFT;
-    double orthoY = (y/double(WINDOW_SIZE_HEIGHT) * ORTHO_HEIGHT) + BOUND_BOTTOM;
-    return dvec2(orthoX, orthoY);
+vec2 getOrthoCoordsFromWindowCoords(float x, float y) {
+    float orthoX = (x/float(WINDOW_SIZE_WIDTH) * ORTHO_WIDTH) + BOUND_LEFT;
+    float orthoY = (y/float(WINDOW_SIZE_HEIGHT) * ORTHO_HEIGHT) + BOUND_BOTTOM;
+    return vec2(orthoX, orthoY);
 }
 vec4 getColorFromVelocity(float v) {
     if (v == 0.0) return vec4(0.0,0.0,0.0,1.0);
@@ -68,7 +68,7 @@ vec4 getColorFromVelocity(float v) {
     return retcolor;
 }
 void main() {
-    dvec2 orthoCoords = getOrthoCoordsFromWindowCoords(double(gl_FragCoord.x), double(gl_FragCoord.y));
+    vec2 orthoCoords = getOrthoCoordsFromWindowCoords(float(gl_FragCoord.x), float(gl_FragCoord.y));
     float normalizedVelocity = findEscapeVelocity(orthoCoords);
 	fragColor = getColorFromVelocity(normalizedVelocity);
 }
